First draft of SWV peak finder
