+++ methodscripts/requirements.txt
@@ -1,3 +1,4 @@
 matplotlib >= 3.5.0
 numpy >= 1.21.1
 pyserial >= 3.5
+scipy >= 1.7.1
